JAVA-41259 Moved code of article jpa-java-time from java-jpa to java-jpa-4
