+++ persistence-modules/java-jpa-4/README.md
@@ -12,3 +12,4 @@ This module contains articles about the Java Persistence API (JPA) in Java.
 - [JPA @Basic Annotation](https://www.baeldung.com/jpa-basic-annotation)
 - [Fixing the JPA error “java.lang.String cannot be cast to Ljava.lang.String;”](https://www.baeldung.com/jpa-error-java-lang-string-cannot-be-cast)
 - [Converting Between LocalDate and SQL Date](https://www.baeldung.com/java-convert-localdate-sql-date)
+- [JPA Support for java.time Types](https://www.baeldung.com/jpa-java-time)
