+++ persistence-modules/java-jpa-4/src/main/resources/META-INF/persistence.xml
@@ -226,5 +226,24 @@
 		</properties>
 	</persistence-unit>
 	
+	<persistence-unit name="java8-datetime-postgresql"
+		transaction-type="RESOURCE_LOCAL">
+		<provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
+		<class>com.baeldung.jpa.datetime.JPA22DateTimeEntity</class>
+		<exclude-unlisted-classes>true</exclude-unlisted-classes>
+		<properties>
+			<property name="jakarta.persistence.jdbc.driver" value="org.postgresql.Driver" />
+			<property name="jakarta.persistence.jdbc.url" value="jdbc:postgresql://localhost:5432/java8-datetime2" />
+			<property name="jakarta.persistence.jdbc.user" value="postgres" />
+			<property name="jakarta.persistence.jdbc.password" value="postgres" />
+			<property name="jakarta.persistence.schema-generation.database.action" value="drop-and-create" />
+
+			<!-- configure logging -->
+			<property name="eclipselink.logging.level" value="INFO" />
+			<property name="eclipselink.logging.level.sql" value="FINE" />
+			<property name="eclipselink.logging.parameters" value="true" />
+		</properties>
+	</persistence-unit>
+	
 	
 </persistence>
