+++ persistence-modules/java-jpa/src/main/resources/META-INF/persistence.xml
@@ -6,7 +6,6 @@
 
 	<persistence-unit name="jpa-h2">
 		<provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-		<class>com.baeldung.jpa.stringcast.Message</class>
 		<class>com.baeldung.jpa.enums.Article</class>
 		<class>com.baeldung.jpa.enums.CategoryConverter</class>
 		<exclude-unlisted-classes>true</exclude-unlisted-classes>
@@ -33,25 +32,6 @@
 			<property name="jakarta.persistence.jdbc.url" value="jdbc:h2:mem:entitygraphdb;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;NON_KEYWORDS=USER" />
 			<property name="jakarta.persistence.schema-generation.database.action" value="drop-and-create" />
 			<property name="jakarta.persistence.sql-load-script-source" value="data-init.sql" />
-		</properties>
-	</persistence-unit>
-
-	<persistence-unit name="java8-datetime-postgresql"
-		transaction-type="RESOURCE_LOCAL">
-		<provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-		<class>com.baeldung.jpa.datetime.JPA22DateTimeEntity</class>
-		<exclude-unlisted-classes>true</exclude-unlisted-classes>
-		<properties>
-			<property name="jakarta.persistence.jdbc.driver" value="org.postgresql.Driver" />
-			<property name="jakarta.persistence.jdbc.url" value="jdbc:postgresql://localhost:5432/java8-datetime2" />
-			<property name="jakarta.persistence.jdbc.user" value="postgres" />
-			<property name="jakarta.persistence.jdbc.password" value="postgres" />
-			<property name="jakarta.persistence.schema-generation.database.action" value="drop-and-create" />
-
-			<!-- configure logging -->
-			<property name="eclipselink.logging.level" value="INFO" />
-			<property name="eclipselink.logging.level.sql" value="FINE" />
-			<property name="eclipselink.logging.parameters" value="true" />
 		</properties>
 	</persistence-unit>
